# middleware/authMiddleware.js
const jwt = require('jsonwebtoken');
const User = require('../models/user');

exports.authenticate = async (req, res, next) => {
  const token = req.headers.authorization?.split(' ')[1];
  if (!token) return res.status(401).json({ message: 'Токен отсутствует' });

  try {
    const decoded = jwt.verify(token, process.env.JWT_SECRET);
    const user = await User.findByPk(decoded.id);

    if (!user) return res.status(404).json({ message: 'Пользователь не найден' });

    req.user = user;
    next();
  } catch (error) {
    return res.status(403).json({ message: 'Недействительный токен' });
  }
};