# swagger/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kinetic SportSystem API",
    "version": "0.1"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_auth_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Read Current User",
        "operationId": "read_current_user_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/auth/update-role": {
      "put": {
        "tags": ["Auth"],
        "summary": "Update Role",
        "operationId": "update_role_auth_update_role_put",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "new_role",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["admin", "hr", "employee"],
              "title": "New Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden: insufficient privileges",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/activities/types": {
      "get": {
        "tags": ["Activities"],
        "summary": "Получить список типов занятий",
        "operationId": "getActivityTypes",
        "responses": {
          "200": {
            "description": "Список типов занятий",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityType"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/activities/types/{id}/classes": {
      "get": {
        "tags": ["Activities"],
        "summary": "Получить классы занятий по типу",
        "operationId": "getActivityClassesByType",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID типа занятия",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список классов занятий",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityClass"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Тип занятия не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/departments": {
      "get": {
        "tags": ["Departments"],
        "summary": "Get all departments",
        "operationId": "getAllDepartments",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of departments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Department"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Departments"],
        "summary": "Create new department",
        "operationId": "createDepartment",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Department created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          }
        }
      }
    },
    "/departments/{id}": {
      "get": {
        "tags": ["Departments"],
        "summary": "Get department by ID",
        "operationId": "getDepartmentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Department details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          },
          "404": {
            "description": "Department not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Departments"],
        "summary": "Update department",
        "operationId": "updateDepartment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Department updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Departments"],
        "summary": "Delete department",
        "operationId": "deleteDepartment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Department deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Department not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/departments/{id}/users": {
      "get": {
        "tags": ["Departments"],
        "summary": "Get users by department ID",
        "operationId": "getUsersByDepartment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID департамента",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Список пользователей департамента",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserShort"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Пользователи не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/events": {
      "get": {
        "tags": ["Events"],
        "summary": "Получить события календаря",
        "operationId": "getEvents",
        "responses": {
          "200": {
            "description": "Список событий",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarEvent"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": ["Events"],
        "summary": "Создать события для пользователей",
        "operationId": "createEventsForUsers",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "События успешно созданы",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CalendarEvent"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/events/{id}": {
      "delete": {
        "tags": ["Events"],
        "summary": "Удалить событие по ID",
        "operationId": "deleteEvent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID события",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Событие удалено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Событие успешно удалено"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Событие не найдено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_login_auth_login_post": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "required": ["username", "password"],
        "title": "Body_login_auth_login_post"
      },
      "HTTPValidationError": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail"
          }
        },
        "title": "HTTPValidationError"
      },
      "Token": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          }
        },
        "required": ["access_token", "token_type"],
        "title": "Token"
      },
      "UserCreate": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" },
          "iin": { "type": "string" },
          "full_name": { "type": "string" },
          "rank": { "type": "string" },
          "role": { "type": "string" },
          "departmentId": { "type": "integer", "nullable": true }
        },
        "required": ["username", "password", "iin", "full_name", "rank", "role"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "full_name": { "type": "string" },
          "rank": { "type": "string" },
          "role": { "type": "string" },
          "departmentId": { "type": "integer", "nullable": true }
        }
      },
      "Department": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "description": { "type": "string" }
        }
      },
      "DepartmentCreate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" }
        },
        "required": ["name"]
      },
      "DepartmentUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "loc": {
            "type": "array",
            "items": {
              "anyOf": [
                { "type": "string" },
                { "type": "integer" }
              ]
            },
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "ActivityType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Сила"
          }
        }
      },
      "ActivityClass": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "11"
          },
          "name": {
            "type": "string",
            "example": "Отжимания"
          }
        }
      },
      "CalendarEvent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 2
          },
          "title": {
            "type": "string",
            "example": "Timed event"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "example": "2025-04-25T12:00:00"
          },
          "location": {
            "type": "string",
            "example": "Astana"
          },
          "type": {
            "type": "string",
            "example": "power"
          },
          "class": {
            "type": "string",
            "example": "push-ups"
          }
        },
        "required": ["title", "start", "location", "type", "class"]
      },
      "UserShort": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 5 },
          "username": { "type": "string", "example": "john_doe" },
          "full_name": { "type": "string", "example": "John Doe" },
          "rank": { "type": "string", "example": "Captain" },
          "role": { "type": "string", "example": "employee" }
        }
      },
      "CreateEventRequest": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "example": "2025-05-01T09:00:00"
          },
          "title": {
            "type": "string",
            "example": "Групповая тренировка"
          },
          "location": {
            "type": "string",
            "example": "Astana"
          },
          "type": {
            "type": "string",
            "example": "Сила"
          },
          "class": {
            "type": "string",
            "example": "Отжимания"
          },
          "userIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "example": 5
            }
          }
        },
        "required": ["start", "title", "location", "type", "class", "userIds"]
      },

      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Тип занятия не найден"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}